Refactor NodeToolCmd classes
