+++ plugins/module_utils/NodeToolCmdObjects.py
@@ -1,0 +1,66 @@
+class NodeToolCmd(object):
+    """
+    This is a generic NodeToolCmd class for building nodetool commands
+    """
+
+    def __init__(self, module):
+        self.module = module
+        self.host = module.params['host']
+        self.port = module.params['port']
+        self.password = module.params['password']
+        self.password_file = module.params['password_file']
+        self.username = module.params['username']
+        self.nodetool_path = module.params['nodetool_path']
+        self.debug = module.params['debug']
+        if self.host is None:
+            self.host = socket.getfqdn()
+
+    def execute_command(self, cmd):
+        return self.module.run_command(cmd)
+
+    def nodetool_cmd(self, sub_command):
+        if self.nodetool_path is not None and len(self.nodetool_path) > 0:
+            if not self.nodetool_path.endswith('/'):  # replace with os.path.join
+                self.nodetool_path += '/'
+        else:
+            self.nodetool_path = ""
+        cmd = "{0}nodetool --host {1} --port {2}".format(self.nodetool_path,
+                                                         self.host,
+                                                         self.port)
+        if self.username is not None:
+            cmd += " --username {0}".format(self.username)
+            if self.password_file is not None:
+                cmd += " --password-file {0}".format(self.password_file)
+            else:
+                cmd += " --password '{0}'".format(self.password)
+        # The thing we want nodetool to execute
+        cmd += " {0}".format(sub_command)
+        if self.debug:
+            self.module.debug(cmd)
+        return self.execute_command(cmd)
+
+
+class NodeTool3PairCommand(NodeToolCmd):
+
+    """
+    Inherits from the NodeToolCmd class. Adds the following methods;
+
+        - status_command
+        - enable_command
+        - disable_command
+    """
+
+    def __init__(self, module, status_cmd, enable_cmd, disable_cmd):
+        NodeToolCmd.__init__(self, module)
+        self.status_cmd = status_cmd
+        self.enable_cmd = enable_cmd
+        self.disable_cmd = disable_cmd
+
+    def status_command(self):
+        return self.nodetool_cmd(self.status_cmd)
+
+    def enable_command(self):
+        return self.nodetool_cmd(self.enable_cmd)
+
+    def disable_command(self):
+        return self.nodetool_cmd(self.disable_cmd)
